requirements.txt
azure.identity
azure.mgmt.containerservice
python_dotenv
flask
opencensus-ext-azure